clone $appRequest before change
had problem with CrudRoute in nette router. While CrudRoute was called first and the basic nette route next, parameter action was always overriden to 'default'. I have traced the stack to ResourceRoute->constructUrl() where it modified Request by calling Request->setParameter(...array('action' => 'default')...). I have cloned the Request so it is only changed for purpose of ResourceRoute and will not change the original Request for purpose of other Routers.